requirements.txt
click==7.1.2
isodate==0.6.0
Jinja2==2.11.2
markdown2==2.3.9
MarkupSafe==1.1.1
pathvalidate==2.3.0
pika==1.1.0
pymongo==3.10.1
pyparsing==2.4.7
python-slugify==4.0.0
PyYAML==5.3.1
rdflib==5.0.0
rdflib-jsonld==0.5.0
six==1.14.0
tenacity==6.2.0
text-unidecode==1.3
